fix delete tools, fix adding tools table

+++ application/models/Tools_m.php
@@ -37,8 +37,18 @@ class Tools_m extends CI_Model {
 		$this->db->update('tools', $params);
 	}
 
-	public function del($id)
-	{
+	public function del($id){
+		$this->db->from('transaksi_detail_peminjaman');
+		$this->db->where('alat_id', $id);
+		$trans = $this->db->get();
+		$this->db->from('transaksi_detail_pengembalian');
+		$this->db->where('alat_id', $id);
+		$trans2 = $this->db->get();
+		if (count($trans->result()) > 0 || count($trans2->result()) > 0 ) {
+			$this->session->set_flashdata('error','gagal menghapus tools <br> <small> tools sedang di gunakan dalam transaksi</small>');
+			redirect('Tools');
+		}
+
 		$this->db->where('alat_id', $id);
 		$this->db->delete('tools');
 	}
